Move risk acknowledgement component to organisms
