+++ packages/app/src/features/dialogs/savings/deposit/views/SavingsDepositView.tsx
@@ -5,10 +5,10 @@ import { DialogActionsPanel } from '@/features/dialogs/common/components/DialogA
 import { FormAndOverviewWrapper } from '@/features/dialogs/common/components/FormAndOverviewWrapper'
 import { MultiPanelDialog } from '@/features/dialogs/common/components/MultiPanelDialog'
 import { DialogForm } from '@/features/dialogs/common/components/form/DialogForm'
-import { RiskAcknowledgement } from '@/features/dialogs/common/components/risk-acknowledgement/RiskAcknowledgement'
 import { AssetInputSchema } from '@/features/dialogs/common/logic/form'
 import { FormFieldsForDialog, PageStatus } from '@/features/dialogs/common/types'
 import { DialogTitle } from '@/ui/atoms/dialog/Dialog'
+import { RiskAcknowledgement } from '@/ui/organisms/risk-acknowledgement/RiskAcknowledgement'
 import { UseFormReturn } from 'react-hook-form'
 import { LiFiTransactionOverview } from '../../common/components/LiFiTransactionOverview'
 import { MakerTransactionOverview } from '../../common/components/maker-transaction-overview'
